+++ dotnet/corecomponents/net.atos.daf.ct2.reports/entity/TripDetails.cs
@@ -14,8 +14,8 @@ namespace net.atos.daf.ct2.reports.entity
         public int Distance { get; set; }
         public int IdleDuration { get; set; }
 
-        public int AverageSpeed { get; set; }
-        public int AverageWeight { get; set; }
+        public double AverageSpeed { get; set; }
+        public double AverageWeight { get; set; }
 
         public long Odometer { get; set; }
 
@@ -23,7 +23,7 @@ namespace net.atos.daf.ct2.reports.entity
         public string EndPosition { get; set; }
 
         public double FuelConsumed { get; set; }
-        public int DrivingTime { get; set; }
+        public long DrivingTime { get; set; }
 
         public int Alert { get; set; }
         public int Events { get; set; }
